Adjust alignement and rendering issues
It seems that manually adjusting the vertical alignment is no longer
needed, since inline-block is doing the correct thing already.

For exporting to PDF a higher resolution image is needed. In theory this
could also be used in the browser, but on "normal" displays that looks
even shittier. The proper approach would probably be to render the image
in multiple size and use a source set...

Ideally the underlying library would need to be rewritten to use SVG.

Addresses splitbrain/dokuwiki-plugin-dw2pdf#433, #14 and #15